Update container Update til docker container. Fungerer, men er meget langsom. Vi eksperimenterer.
