+++ .dockerignore
@@ -1,0 +1,13 @@
+.dockerignore
+
+# Node
+node_modules
+npm-debug.log
+dist
+
+# Git
+.git
+.gitignore
+
+# VS Code
+.vscode
