--- green-house-app/src/routes/Hjem.js
@@ -1,5 +1,0 @@
-export default function Hjem() {
-    return <> 
-    <h1 id="frontpage-greeting">Velkommen</h1>
-</>
-}
