+++ public/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Green house</title>
+  </head>
+  <body>
+    <div id="root"></div>
+  </body>
+</html>
