+++ src/routes/Hjem.js
@@ -1,0 +1,5 @@
+export default function Hjem() {
+    return <> 
+    <h1 id="frontpage-greeting">Velkommen</h1>
+</>
+}
